[NFC] Factor out actor isolation inference tests into subdirectory
