+++ test/Concurrency/ActorIsolationInference/Inputs/other_global_actor.swift
