+++ test/Concurrency/ActorIsolationInference/global_actor_swift6.swift
@@ -1,12 +1,12 @@
 // RUN: %empty-directory(%t)
 
-// RUN: %target-swift-frontend -swift-version 6 -emit-module -emit-module-path %t/other_global_actor_inference.swiftmodule -module-name other_global_actor_inference -strict-concurrency=complete %S/Inputs/other_global_actor_inference.swift
+// RUN: %target-swift-frontend -swift-version 6 -emit-module -emit-module-path %t/other_global_actor.swiftmodule -module-name other_global_actor -strict-concurrency=complete %S/Inputs/other_global_actor.swift
 
 // RUN: %target-swift-frontend -swift-version 6 -I %t -disable-availability-checking %s -emit-sil -o /dev/null -verify
 
 // REQUIRES: concurrency
 
-import other_global_actor_inference
+import other_global_actor
 
 actor SomeActor { }
